+++ ideas/src/app/core/index.ts
@@ -1,6 +1,6 @@
 import { ObjectId } from 'mongodb'
 import sanitizeHtml from 'sanitize-html'
-import { ScoreNames } from '../types'
+import { ScoreNames } from './score.types'
 import { buildMakeIdea } from './idea'
 import { makeScore, getScore } from './score'
 
@@ -14,5 +14,10 @@ const makeScores = () => {
   return Object.values(ScoreNames).map(type => makeScore({ type }))
 }
 const makeIdea = buildMakeIdea({ sanitize, Id, makeScores, getScore })
+
+export * from './idea.types'
+export * from './score.types'
+export * from './idea'
+export * from './score'
 
 export { Id, makeScores, makeIdea, makeScore }
