+++ ideas/src/app/core/score.types.ts
