--- ideas/src/app/core/types/Immutable.ts
@@ -1,5 +1,0 @@
-type Immutable<T> = {
-  readonly [K in keyof T]: Immutable<T[K]>
-}
-
-export { Immutable }
