--- ideas/src/app/core/types/index.ts
@@ -1,3 +1,0 @@
-export * from './IdeaTypes'
-export * from './ScoreTypes'
-export * from './Immutable'
