+++ ideas/src/app/useCase.types.ts
